+++ runner/testdata/test-build/sum_test.go
@@ -1,0 +1,8 @@
+package sum
+
+import "testing"
+
+func TestAdd(t *testing.T) {
+	var total string = Add(1, 2)
+	_ = total
+}