+++ runner/testdata/test-race/counter_test.go
@@ -1,0 +1,7 @@
+package counter
+
+import "testing"
+
+func TestCount(t *testing.T) {
+	Count()
+}